fix: import BaseProvider from legacy alias
